Added the simplest deinterlacer implementation
The application takes two arguments, the paths to input and output pictures, both JPEG.
Build system is CMake because it takes the possibility to be easy
converted to others.
Current implementation uses prinstalled  libjpeg - in the next update
would be changed to thirdparty internal project folder.
Processing must be done in the YCbCr color space.

“Blending” algorithm: line N in the output image is the average of lines N and N-1 from the input image. The first line stays unchanged.